rudimentary fix for the problem in hub2
When going back from a level in hub2, it took you over to some random pesky coordinates for some odd reason and so I made a rudimentary fix which involves making an object, when, in collision with the player, warps the player to a set of coordinates. It's not a permanent fix, but at least you aren't ending up in the void anymore!