Retro Madness now considered "Level 5" and Level 5 now has working health
Extra changes: aspect ratio is now locked at 16:9 again. It was fun to see the game ultrawide but unfortunately because the game was not made for it there are tons of bugs with it. If you want to fix it though feel free to help me and make a pull request!